CLN/TST: move ops test from test_frame.py to test_expressions.py
